+++ format.php
@@ -35,23 +35,36 @@ if ($topic = optional_param('topic', 0, PARAM_INT)) {
     redirect($url);
 }
 
+// Retrieve course format option fields and add them to the $course object.
+$format = core_courseformat\base::instance($course);
+$course = $format->get_course();
 $context = context_course::instance($course->id);
 
+// Add any extra logic here.
 if (($marker >= 0) && has_capability('moodle/course:setcurrentsection', $context) && confirm_sesskey()) {
     $course->marker = $marker;
     course_set_marker($course->id, $marker);
 }
 
-$course = course_get_format($course)->get_course();
-
+// Make sure section 0 is created.
 course_create_sections_if_missing($course, range(0, $course->numsections));
 
-$renderer = $PAGE->get_renderer('format_buttons');
+$renderer = $format->get_renderer($PAGE);
 
+// Setup the format base instance.
 if (!empty($displaysection)) {
-    $renderer->print_single_section_page($course, null, null, null, null, $displaysection);
-} else {
+    // $renderer->print_single_section_page($course, null, null, null, null, $displaysection);
     $renderer->print_multiple_section_page($course, null, null, null, null);
+    $format->set_section_number($displaysection);
 }
+
+// $renderer->print_single_section_page($course, null, null, null, null, $displaysection);
+// $renderer->print_multiple_section_page($course, null, null, null, null);
+// $format->set_section_number($displaysection);
+// Output course content.
+$outputclass = $format->get_output_classname('content');
+// mtrace('CLS '.$outputclass);
+$widget = new $outputclass($format);
+echo $renderer->render($widget);
 
 $PAGE->requires->js('/course/format/buttons/format.js');
